Perf: world-gen optimizations + working-tree checkpoint
Landed this session (safe, accuracy-preserving, verified compiling):
- McWorld/McCoordinator: load-phase budget boost (loadPhaseUpdateBudgetMs)
  raises per-frame budgets during the one-time initial gen, auto-restores after.
- McWorld: GPU_FACE_READBACKS_PER_FRAME 2 -> 8 (drain GPU-meshed chunks faster).
- Scene: profiling instrumentation flags disabled (removes per-frame Udon overhead).
- OPTIMIZATION_NOTES.md: profiling diagnosis (gen is CPU/serialization-bound,
  GPU ~92% idle) + staged plans for instanced draw, generator parallelization,
  RT memory.

Also checkpoints prior untracked working state (GPU shaders, GPU_Neo materials,
Roslyn plugins, notes). Excludes debug screenshots.

+++ Assets/Editor/VRCMinecraft/McTestBullshit.cs
@@ -46,7 +46,7 @@ public class SeedTesterWindow : EditorWindow
         // Button to trigger the seed calculation
         if (GUILayout.Button("Calculate Seed"))
         {
-            int calculatedSeed = McUtils.GetMinecraftSeed(seedInput);
+            long calculatedSeed = McUtils.GetMinecraftSeed(seedInput);
             seedResultText = $"Seed Result: {calculatedSeed}";
             permTableResultText = "(Permutation table cleared)";
         }
@@ -54,9 +54,9 @@ public class SeedTesterWindow : EditorWindow
         // Button to trigger permutation table generation
         if (GUILayout.Button("Generate Permutation Table"))
         {
-            int seed = McUtils.GetMinecraftSeed(seedInput);
+            long seed = McUtils.GetMinecraftSeed(seedInput);
             seedResultText = $"Using Seed: {seed}";
-            
+
             int[] permTable = McUtils.GetPermutationTable(new JavaRandom(seed));
             
             // Format the byte array into a readable string
